+++ ximpol/examples/J1708.py
@@ -28,21 +28,23 @@ from ximpol.evt.event import xEventFile
 from ximpol.utils.matplotlib_ import pyplot as plt
 from ximpol.utils.matplotlib_ import save_current_figure
 from ximpol.evt.binning import xBinnedPhasogram
-from ximpol.config.gk_per import phasogram_spline, pol_degree_spline,\
-    pol_angle_spline
+from ximpol.config.J1708 import energy_spectrum, polarization_degree,\
+    polarization_angle
+from ximpol.evt.event import xEventFile
+from ximpol.irf import load_mrf
 
 
 """Script-wide simulation and analysis settings.
 """
-CFG_FILE_PATH = os.path.join(XIMPOL_CONFIG, 'gk_per.py')
-OUT_FILE_PATH_BASE = os.path.join(XIMPOL_DATA, 'gk_per')
+CFG_FILE_PATH = os.path.join(XIMPOL_CONFIG, 'J1708.py')
+OUT_FILE_PATH_BASE = os.path.join(XIMPOL_DATA, 'J1708')
 EVT_FILE_PATH = '%s.fits' % OUT_FILE_PATH_BASE
 ANALYSIS_FILE_PATH = '%s_analysis.txt' % OUT_FILE_PATH_BASE
-SIM_DURATION = 800000.
-#PHASE_BINNING = numpy.linspace(0., 1., 6)
-PHASE_BINNING = numpy.array([0., 0.30, 0.40, 0.50, 0.60, 0.70, 1.])
+SIM_DURATION = 250000.
+PHASE_BINNING = numpy.linspace(0., 1., 10)
+#PHASE_BINNING = numpy.array([0., 0.30, 0.40, 0.50, 0.60, 0.70, 1.])
 PHASE_BINNING = zip(PHASE_BINNING[:-1], PHASE_BINNING[1:])
-E_BINNING = [2., 4., 8.]
+E_BINNING = [2., 6]
 
 
 def _sel_file_path(i):
@@ -75,7 +77,7 @@ def generate():
     """Generate the events.
     """
     PIPELINE.xpobssim(configfile=CFG_FILE_PATH, duration=SIM_DURATION,
-                      outfile=EVT_FILE_PATH, seed=2)
+                      outfile=EVT_FILE_PATH)
 
 def bin():
     """Bin the event file in different flavors for the actual analysis.
@@ -88,7 +90,7 @@ def bin():
     for i, (_emin, _emax) in enumerate(_ebinning):
         PIPELINE.xpselect(EVT_FILE_PATH, emin=_emin, emax=_emax,
                           outfile=_esel_file_path(i))
-        PIPELINE.xpbin(_esel_file_path(i), algorithm='PHASG', phasebins=10)
+        PIPELINE.xpbin(_esel_file_path(i), algorithm='PHASG', phasebins=9 )
     for i, (_min, _max) in enumerate(PHASE_BINNING):
         PIPELINE.xpselect(EVT_FILE_PATH, phasemin=_min, phasemax=_max,
                           outfile=_sel_file_path(i))
@@ -102,13 +104,22 @@ def analyze():
         logger.info('%s exists, delete it if you want to recreate it.' %\
                     ANALYSIS_FILE_PATH)
         return
+    modf = load_mrf('xipe_goal')
     logger.info('Opening output file %s...' % ANALYSIS_FILE_PATH)
     analysis_file = open(ANALYSIS_FILE_PATH, 'w')
     for i, (_min, _max) in enumerate(PHASE_BINNING):
+        _evt_file = xEventFile(_sel_file_path(i))
+        _energy = _evt_file.event_data['ENERGY']
+        _phase = _evt_file.event_data['PHASE']
+        exp = (polarization_degree(_energy, _phase, 0, 0)*modf(_energy)).sum()/\
+              modf(_energy).sum()
         _mcube = xBinnedModulationCube(_mcube_file_path(i))
         _mcube.fit()
         _fit_results = _mcube.fit_results[-1]
         print _fit_results
+        print exp
+        print polarization_degree(_mcube.emean[-1], _phase, 0, 0)
+        raw_input()
         _phase = 0.5*(_min + _max)
         _phase_err = 0.5*(_max - _min)
         _pol_deg = _fit_results.polarization_degree
@@ -130,6 +141,8 @@ def plot(save_plots=False):
     mod_label = 'Input model'
     _phase, _phase_err, _pol_deg, _pol_deg_err, _pol_angle,\
         _pol_angle_err = numpy.loadtxt(ANALYSIS_FILE_PATH, unpack=True)
+    _pol_angle = numpy.degrees(_pol_angle)
+    _pol_angle_err = numpy.degrees(_pol_angle_err)
     _colors = ['blue']*len(_pol_deg)
     plt.figure('Polarization degree')
     _good_fit = _pol_deg > 2*_pol_deg_err
@@ -140,13 +153,16 @@ def plot(save_plots=False):
     plt.errorbar(_phase[_bad_fit], _pol_deg[_bad_fit],
                  xerr=_phase_err[_bad_fit], yerr=_pol_deg_err[_bad_fit],
                  fmt='o', color='gray')
-    pol_degree_spline.plot(show=False, label=mod_label, color='green')
-    plt.axis([0., 1., 0., 0.1])
+    #pol_degree_spline.plot(show=False, label=mod_label, color='green')
+    _x = numpy.linspace(0, 1, 100)
+    _y = polarization_degree(2.38, _x, 0, 0)
+    plt.plot(_x, _y)
+    plt.axis([0., 1., 0., 1.])
     plt.legend(bbox_to_anchor=(0.37, 0.95))
     plt.figtext(0.6, 0.8, '%.2f--%.2f keV' %\
                 (E_BINNING[0], E_BINNING[-1]), size=16)
     if save_plots:
-        plt.savefig('gk_per_polarization_degree.png')
+        plt.savefig('J1708_per_polarization_degree.png')
     plt.figure('Polarization angle')
     plt.errorbar(_phase[_good_fit], _pol_angle[_good_fit],
                  xerr=_phase_err[_good_fit], yerr=_pol_angle_err[_good_fit],
@@ -154,31 +170,35 @@ def plot(save_plots=False):
     plt.errorbar(_phase[_bad_fit], _pol_angle[_bad_fit],
                  xerr=_phase_err[_bad_fit], yerr=_pol_angle_err[_bad_fit],
                  fmt='o', color='gray')
-    pol_angle_spline.plot(show=False, label=mod_label, color='green',
-                          scale=numpy.radians(1.))
-    plt.axis([0., 1., -0.1, 1.5])
+    #pol_angle_spline.plot(show=False, label=mod_label, color='green',
+    #                      scale=numpy.radians(1.))
+    _x = numpy.linspace(0, 1, 100)
+    _y = numpy.degrees(polarization_angle(2.56, _x, 0, 0))
+    plt.plot(_x, _y)
+    plt.axis([0., 1., 0, 180])
     plt.xlabel('Rotational phase')
-    plt.ylabel('Polarization angle [rad]')
+    plt.ylabel('Polarization angle [$^\\circ$]')
     plt.legend(bbox_to_anchor=(0.37, 0.95))
     plt.figtext(0.6, 0.8, '%.2f--%.2f keV' %\
                 (E_BINNING[0], E_BINNING[-1]), size=16)
     if save_plots:
-        plt.savefig('gk_per_polarization_angle.png')
+        plt.savefig('J1708_per_polarization_angle.png')
     _ebinning = zip(E_BINNING[:-1], E_BINNING[1:])
     if len(_ebinning) > 1:
         _ebinning.append((E_BINNING[0], E_BINNING[-1]))
     for i, (_emin, _emax) in enumerate(_ebinning):
         plt.figure('Phasogram %d' % i)
         phasogram = xBinnedPhasogram(_phasg_file_path(i))
-        _scale = phasogram.counts.sum()/phasogram_spline.norm()/\
-                 len(phasogram.counts)
-        phasogram_spline.plot(show=False, label=mod_label, scale=_scale,
-                              color='green')
+        #_scale = phasogram.counts.sum()/phasogram_spline.norm()/\
+        #         len(phasogram.counts)
+        #phasogram_spline.plot(show=False, label=mod_label, scale=_scale,
+        #                      color='green')
+        #print phasogram.counts
         phasogram.plot(show=False, color='blue', label=sim_label )
         plt.legend(bbox_to_anchor=(0.37, 0.95))
         plt.figtext(0.65, 0.8, '%.2f--%.2f keV' % (_emin, _emax), size=16)
         if save_plots:
-            plt.savefig('gk_per_phasogram_%d.png' % i)
+            plt.savefig('J1708_per_phasogram_%d.png' % i)
     plt.show()
 
 def run(save_plots):